Refactroing code according to rubocop
