+++ app.rb
@@ -4,9 +4,11 @@ require_relative 'book'
 require_relative 'rental'
 require_relative 'teacher'
 require_relative 'classroom'
+require './ui_creater'
 
-class App
+class App < UiCreater
   def initialize
+    super()
     @classroom = []
     @books = []
     @rentals = []
@@ -17,24 +19,10 @@ class App
     ui_creater('THANKS FOR USING OUR APPLICATION')
   end
 
-  def ui_creater(prop)
-    puts "-----------------------------------------\n#{prop} \n-----------------------------------------"
-  end
-
-  def title_ui(title = '')
-    puts "******* Select an Option #{title} *******" 
-  end
-
-  def table_ui(table_title, line)
-    puts "========================== #{table_title} =========================="
-    puts line
-    puts "========================== *************** =========================="
-  end
-
   def list_all_books
     if @books.length.positive?
       title = 'Book Shelf'
-      line = ""
+      line = ''
       @books.each do |book|
         line += "#{book.title.upcase} by #{book.author.upcase} \n"
       end
@@ -46,14 +34,14 @@ class App
 
   def list_all_person
     if @person.length.positive?
-      title = "List of persons"
-      lines = ""
+      title = 'List of persons'
+      lines = ''
       @person.each do |person|
-        if person.respond_to?('specialization')
-          lines += "#{person.name.upcase}, he is #{person.age} years old, specilites in #{person.specialization}\n"
-        else
-          lines += "#{person.name.upcase}, he is #{person.age} years old, study in #{person.classroom.label} \n"
-        end
+        lines += if person.respond_to?('specialization')
+                   "#{person.name.upcase}, he is #{person.age} years old, specilites in #{person.specialization}\n"
+                 else
+                   "#{person.name.upcase}, he is #{person.age} years old, study in #{person.classroom.label} \n"
+                 end
       end
       table_ui(title, lines)
     else
@@ -62,7 +50,7 @@ class App
   end
 
   def create_a_person
-  title_ui
+    title_ui
     puts '1. To create a student'
     puts '2. To create a teacher'
     print 'Enter the option : '
@@ -108,6 +96,24 @@ class App
     ui_creater("created book successfully with title of #{new_book.title.upcase} authored by #{new_book.author.upcase}")
   end
 
+  def book_table
+    title1 = 'Book Shelf'
+    line1 = ''
+    @books.each_with_index do |book, index|
+      line1 += "#{index} Title #{book.title}, authored by #{book.author} \n"
+    end
+    table_ui(title1, line1)
+  end
+
+  def person_table
+    title2 = 'Persons'
+    line2 = ''
+    @person.each_with_index do |person, index|
+      line2 += "#{index} #{person.name.upcase} with id #{person.id} and #{person.age} years old \n"
+    end
+    table_ui(title2, line2)
+  end
+
   def create_rental
     if @books.empty? || @person.empty?
       return ui_creater('There are no books available or registered person at the moment for rental')
@@ -115,20 +121,10 @@ class App
 
     puts 'Select a book from the shelf: '
     puts ''
-    title1 = "Book Shelf"
-    line1 = ""
-    @books.each_with_index do |book, index|
-      line1 += "#{index} Title #{book.title}, authored by #{book.author} \n"
-    end
-    table_ui(title1,line1)
+    book_table
     choice_book = @books[gets.chomp.to_i]
     puts 'Select a person to rent the book by id: '
-    title2 = "Persons"
-    line2 = ""
-    @person.each_with_index do |person, index|
-      line2 += "#{index} #{person.name.upcase} with id #{person.id} and #{person.age} years old \n"
-    end
-    table_ui(title2, line2)
+    person_table
     choice_person = @person[gets.chomp.to_i]
     print 'Enter date of rental to keep track of book : '
     date = gets.chomp
@@ -144,9 +140,9 @@ class App
   end
 
   def display_persons
-    title_ui("i.e. person id")
-    title = "Person"
-    line = ""
+    title_ui('i.e. person id')
+    title = 'Person'
+    line = ''
     @person.each { |person| line += "ID: #{person.id}, Name: #{person.name.upcase} and Age #{person.age} \n" }
     table_ui(title, line)
   end
@@ -162,7 +158,7 @@ class App
 
   def display_rentals_by_person(person)
     title = "rentals by #{person.name.upcase}"
-    line =""
+    line = ''
     if person&.rentals&.any?
       person.rentals.each do |item|
         line += "Book: #{item.book.title.upcase} authored by #{item.book.author.upcase} on the date #{item.date} \n"
